+++ CHANGELOG.md
@@ -18,6 +18,8 @@ All notable changes of GetSimple Legacy CMS will be documented in this file.
 - Added theme function `get_page_summary()` to echo or return the page summary.
 - Added the field *Page Featured Image* to store URL of featured image of the page.
 - Added theme function `get_page_featured_image()` to echo or return the URL of the page featured image.
+- Added theme function `get_component_title()` to echo or return title of a component.
+- Added theme function `get_component_description()` to echo or return description of a component.
 
 ## [2025.1.1] - 2025.01.28